Moved weather utils to manager.
