+++ src/engine/core/managers/weather/utils/weather_generic.ts
